Move 1Password secret fetch helpers to support/op-utils.js and update Cypress config to use them
