+++ test/e2e/cypress.config.js
@@ -2,7 +2,7 @@ const { defineConfig } = require("cypress");
 const { authenticateOAuthDevice } = require("./support/oauth-task");
 const fs = require("fs");
 const path = require("path");
-const { execSync } = require("child_process");
+const { get1PasswordSecret } = require("./support/op-utils");
 
 const DEBUG_CYPRESS = process.env.DEBUG_CYPRESS === "true";
 const ACTIONS_STEP_DEBUG = process.env.ACTIONS_STEP_DEBUG === "true";
@@ -12,49 +12,11 @@ const isCI = process.env.CI === "true";
 const configPath = path.resolve(__dirname, "config/staging-pccqa.json");
 const pccConfig = JSON.parse(fs.readFileSync(configPath, "utf8"));
 
-// 1Password CLI secret fetch helper
-function get1PasswordSecret(item, field, vault) {
-  try {
-    // Prefer system PATH op, fallback to local bin/op
-    function findOpInPath() {
-      try {
-        const which = process.platform === "win32" ? "where" : "which";
-        const opPath = execSync(`${which} op`, { encoding: "utf8" })
-          .split(/\r?\n/)[0]
-          .trim();
-        if (opPath && fs.existsSync(opPath)) {
-          return opPath;
-        }
-      } catch {
-        // ignore errors, treat as not found
-      }
-      return null;
-    }
-    const localOp = path.resolve(
-      __dirname,
-      "../bin",
-      process.platform === "win32" ? "op.exe" : "op",
-    );
-    const opCommand =
-      findOpInPath() || (fs.existsSync(localOp) ? localOp : "op");
-    const result = execSync(
-      `"${opCommand}" item get "${item}" --field "${field}" --vault "${vault}" --reveal`,
-      { encoding: "utf8" },
-    );
-    return result.trim();
-  } catch (error) {
-    console.warn(
-      `Warning: Could not fetch 1Password secret '${item}': ${error.message}`,
-    );
-    return "PLACEHOLDER_PASSWORD";
-  }
-}
-
 // Rewrite placeholder password with actual secret and handle failures
 if (process.env.CI === "true" && process.env.PCC_USER_CCQA3) {
   pccConfig.pcc_user_ccqa3.auth.password = process.env.PCC_USER_CCQA3;
 } else if (pccConfig.pcc_user_ccqa3.auth.password === "UPDATE") {
-  // Replace with your actual item, field, and vault names
+  // Update as needed with correct 1pass vault and item names
   pccConfig.pcc_user_ccqa3.auth.password = get1PasswordSecret(
     "pcc_user_ccqa3",
     "password",
